reorganized folders / added definition file with header guard
